+++ lib/cuprum/collections/repository.rb
@@ -84,10 +84,10 @@ module Cuprum::Collections
     end
     alias << add
 
-    # @overload create(collection_name: nil, entity_class: nil, force: false, **options)
+    # @overload create(name: nil, entity_class: nil, force: false, **options)
     #   Adds a new collection with the given name to the repository.
     #
-    #   @param collection_name [String] the name of the new collection.
+    #   @param name [String] the name of the new collection.
     #   @param entity_class [Class, String] the class of entity represented in
     #     the collection.
     #   @param force [true, false] if true, override an existing collection with
@@ -106,6 +106,23 @@ module Cuprum::Collections
       collection
     end
 
+    # @overload find(name: nil, entity_class: nil, **options)
+    def find(**parameters)
+      qualified_name = qualified_name_for(**parameters)
+      collection     = @collections[qualified_name]
+
+      unless collection
+        raise UndefinedCollectionError,
+          "repository does not define collection #{qualified_name.inspect}"
+      end
+
+      return collection if collection.matches?(**parameters)
+
+      error_message = partial_match_error_message(collection:, parameters:)
+
+      raise DuplicateCollectionError, error_message
+    end
+
     # @overload find_or_create(collection_name: nil, entity_class: nil, **options)
     #   Finds or creates a new collection with the given name.
     #
@@ -118,7 +135,12 @@ module Cuprum::Collections
     #
     #   @raise [DuplicateCollectionError] if a collection with the same name
     #     but different parameters already exists in the repository.
-    def find_or_create(**parameters)
+    def find_or_create(**parameters) # rubocop:disable Metrics/MethodLength
+      tools.core_tools.deprecate(
+        "#{self.class.name}#find_or_create()",
+        message: 'Use #create or #find method.'
+      )
+
       qualified_name = qualified_name_for(**parameters)
 
       unless key?(qualified_name)
@@ -144,6 +166,20 @@ module Cuprum::Collections
       @collections.key?(qualified_name.to_s)
     end
 
+    # Removes the specified collection from the repository.
+    #
+    # @param qualified_name [String, Symbol] the name of the collection to
+    #   remove.
+    #
+    # @return [Cuprum::Collections::Collection] the removed collection.
+    def remove(qualified_name:)
+      collection = find(qualified_name:)
+
+      @collections.delete(collection.qualified_name)
+
+      collection
+    end
+
     private
 
     def build_collection(**)
@@ -152,10 +188,28 @@ module Cuprum::Collections
         'subclass and implement the #build_collection method.'
     end
 
+    def partial_match_error_message(collection:, parameters:)
+      non_matching =
+        parameters
+          .each_key
+          .reject { |key| collection.public_send(key) == parameters[key] }
+      expected = parameters.slice(*non_matching)
+      actual   = non_matching.to_h { |key| [key, collection.public_send(key)] }
+      name     = collection.qualified_name.inspect
+
+      "collection #{name} exists but does not match:\n" \
+        "\n  expected: #{expected.inspect}" \
+        "\n    actual: #{actual.inspect}"
+    end
+
     def qualified_name_for(**parameters)
       Cuprum::Collections::Relations::Parameters
         .resolve_parameters(parameters)
         .fetch(:qualified_name)
+    end
+
+    def tools
+      SleepingKingStudios::Tools::Toolbelt.instance
     end
 
     def valid_collection?(collection)