+++ lib/cuprum/collections/rspec/deferred/repository_examples.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 require 'rspec/sleeping_king_studios/deferred'
+require 'rspec/sleeping_king_studios/matchers/core/have_aliased_method'
 
 require 'cuprum/collections/rspec/deferred'
 
@@ -457,6 +458,161 @@ module Cuprum::Collections::RSpec::Deferred
         end
       end
 
+      describe '#find' do
+        shared_examples 'should find the matching collection' do
+          context 'when the collection does not exist' do
+            let(:error_message) do
+              "repository does not define collection #{qualified_name.inspect}"
+            end
+
+            it 'should raise an exception' do
+              expect { repository.find(**collection_options) }.to raise_error(
+                described_class::UndefinedCollectionError,
+                error_message
+              )
+            end
+          end
+
+          next if deferred_options.fetch(:abstract, false)
+
+          context 'when the collection exists' do
+            include_deferred 'when the repository has many collections'
+
+            let(:collection)     { collections.values.first }
+            let(:entity_class)   { collection.entity_class }
+            let(:name)           { collection.name }
+            let(:qualified_name) { collection.qualified_name }
+
+            it 'should find the collection' do
+              expect(repository.find(**collection_options)).to be == collection
+            end
+          end
+        end
+
+        let(:name)               { 'books' }
+        let(:qualified_name)     { name.to_s }
+        let(:collection_options) { {} }
+
+        it 'should define the method' do
+          expect(repository)
+            .to respond_to(:find)
+            .with(0).arguments
+            .and_any_keywords
+        end
+
+        describe 'with no parameters' do
+          let(:error_message) { "name or entity class can't be blank" }
+
+          it 'should raise an exception' do
+            expect { repository.find }
+              .to raise_error ArgumentError, error_message
+          end
+        end
+
+        if deferred_options.fetch(:find_by_entity_class, true)
+          describe 'with entity_class: a Class' do
+            let(:entity_class)       { Book }
+            let(:collection_options) { super().merge(entity_class:) }
+
+            include_examples 'should find the matching collection'
+          end
+
+          describe 'with entity_class: a String' do
+            let(:entity_class)       { 'Book' }
+            let(:collection_options) { super().merge(entity_class:) }
+
+            include_examples 'should find the matching collection'
+          end
+        end
+
+        describe 'with name: a String' do
+          let(:collection_options) { super().merge(name: name.to_s) }
+
+          include_examples 'should find the matching collection'
+        end
+
+        describe 'with name: a Symbol' do
+          let(:collection_options) { super().merge(name: name.intern) }
+
+          include_examples 'should find the matching collection'
+        end
+
+        describe 'with qualified_name: a String' do
+          let(:collection_options) do
+            super().merge(qualified_name: qualified_name.to_s)
+          end
+
+          include_examples 'should find the matching collection'
+        end
+
+        describe 'with qualified_name: a Symbol' do
+          let(:collection_options) do
+            super().merge(qualified_name: qualified_name.intern)
+          end
+
+          include_examples 'should find the matching collection'
+        end
+
+        describe 'with multiple parameters' do
+          let(:entity_class) { Book }
+          let(:collection_options) do
+            super().merge(entity_class:, qualified_name:)
+          end
+
+          context 'when the collection does not exist' do
+            let(:error_message) do
+              "repository does not define collection #{qualified_name.inspect}"
+            end
+
+            it 'should raise an exception' do
+              expect { repository.find(**collection_options) }.to raise_error(
+                described_class::UndefinedCollectionError,
+                error_message
+              )
+            end
+          end
+
+          next if deferred_options.fetch(:abstract, false)
+
+          context 'when a partially-matching collection exists' do
+            include_deferred 'when the repository has many collections'
+
+            let(:collection)     { collections.values.first }
+            let(:entity_class)   { Spec::EntityClass }
+            let(:qualified_name) { collection.qualified_name }
+            let(:error_message) do
+              <<~TEXT.strip
+                collection "authors" exists but does not match:
+
+                  expected: #{{ entity_class: Spec::EntityClass }.inspect}
+                    actual: #{{ entity_class: Hash }.inspect}
+              TEXT
+            end
+
+            example_class 'Spec::EntityClass'
+
+            it 'should raise an exception' do
+              expect { repository.find(**collection_options) }.to raise_error(
+                described_class::DuplicateCollectionError,
+                error_message
+              )
+            end
+          end
+
+          context 'when a matching collection exists' do
+            include_deferred 'when the repository has many collections'
+
+            let(:collection)     { collections.values.first }
+            let(:entity_class)   { collection.entity_class }
+            let(:qualified_name) { collection.qualified_name }
+
+            it 'should find the collection' do
+              expect(repository.find(**collection_options)).to be == collection
+            end
+          end
+        end
+      end
+
       describe '#find_or_create' do
         let(:collection_name)    { 'books' }
         let(:qualified_name)     { collection_name.to_s }
@@ -473,7 +629,12 @@ module Cuprum::Collections::RSpec::Deferred
             'repository subclass and implement the #build_collection method.'
         end
 
-        def create_collection(safe: true, **options)
+        before(:example) do
+          allow(SleepingKingStudios::Tools::Toolbelt.instance.core_tools)
+            .to receive(:deprecate)
+        end
+
+        define_method :create_collection do |safe: true, **options|
           if safe
             begin
               repository.find_or_create(**collection_options, **options)
@@ -505,6 +666,17 @@ module Cuprum::Collections::RSpec::Deferred
           end
 
           next
+        end
+
+        it 'should print a deprecation warning' do
+          repository.find_or_create(qualified_name:)
+
+          expect(SleepingKingStudios::Tools::Toolbelt.instance.core_tools)
+            .to have_received(:deprecate)
+            .with(
+              'Cuprum::Collections::Basic::Repository#find_or_create()',
+              message: 'Use #create or #find method.'
+            )
         end
 
         describe 'with entity_class: a Class' do
@@ -627,6 +799,90 @@ module Cuprum::Collections::RSpec::Deferred
 
         wrap_deferred 'when the repository has many collections' do
           it { expect(repository.keys).to be == collections.keys }
+        end
+      end
+
+      describe '#remove' do
+        let(:qualified_name) { 'books' }
+
+        it 'should define the method' do
+          expect(repository)
+            .to respond_to(:remove)
+            .with(0).arguments
+            .and_keywords(:qualified_name)
+        end
+
+        describe 'with qualified_name: a String' do
+          let(:qualified_name) { super().to_s }
+
+          context 'when the collection does not exist' do
+            let(:error_message) do
+              "repository does not define collection #{qualified_name.inspect}"
+            end
+
+            it 'should raise an exception' do
+              expect { repository.remove(qualified_name:) }.to raise_error(
+                described_class::UndefinedCollectionError,
+                error_message
+              )
+            end
+          end
+
+          next if deferred_options.fetch(:abstract, false)
+
+          context 'when the collection exists' do
+            include_deferred 'when the repository has many collections'
+
+            let(:collection)     { collections.values.first }
+            let(:qualified_name) { collection.qualified_name.to_s }
+
+            it 'should return the collection' do
+              expect(repository.remove(qualified_name:)).to be == collection
+            end
+
+            it 'should remove the collection from the repository' do
+              expect { repository.remove(qualified_name:) }
+                .to change(repository, :keys)
+                .to(satisfy { |keys| !keys.include?(qualified_name.to_s) })
+            end
+          end
+        end
+
+        describe 'with qualified_name: a Symbol' do
+          let(:qualified_name) { super().intern }
+
+          context 'when the collection does not exist' do
+            let(:error_message) do
+              'repository does not define collection ' \
+                "#{qualified_name.to_s.inspect}"
+            end
+
+            it 'should raise an exception' do
+              expect { repository.remove(qualified_name:) }.to raise_error(
+                described_class::UndefinedCollectionError,
+                error_message
+              )
+            end
+          end
+
+          next if deferred_options.fetch(:abstract, false)
+
+          context 'when the collection exists' do
+            include_deferred 'when the repository has many collections'
+
+            let(:collection)     { collections.values.first }
+            let(:qualified_name) { collection.qualified_name.intern }
+
+            it 'should return the collection' do
+              expect(repository.remove(qualified_name:)).to be == collection
+            end
+
+            it 'should remove the collection from the repository' do
+              expect { repository.remove(qualified_name:) }
+                .to change(repository, :keys)
+                .to(satisfy { |keys| !keys.include?(qualified_name.to_s) })
+            end
+          end
         end
       end
     end